Add pg_collation_index_dependencies and pg_collation_constraint_dependencies SRF
For a given index or constraint, try to find all possible collation this index
depends on.

This is only an initial work, without any kind of intelligence in the way
collations are detected to avoid tracking collations that are not really used
by an index or an expression, so false positive are expected.